pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "eeve"
version = "0.1.0"
description = "Eeve: a toolkit for data processing and training machine translation models"
requires-python = ">=3.10"
dependencies = [
    "transformers<=4.53.3",
    "datatrove[all]",
    "sentence-transformers>=5.1.0",
    "fasttext-numpy2-wheel==0.9.2",
    "hydra-core>=1.3.2",
    "rich",
    "sentencepiece>=0.2.0",
    "datasets>=3.2.0",
    "tqdm",
    "PyYAML>=5.1.0",
    "sacrebleu",
    "evaluate",
    "faiss-cpu"
]

[tool.setuptools.packages.find]
where = ["eeve"] 

[tool.ruff]
lint.ignore = [
    "C901", # `function_name` is too complex
    "E501", # line length violation
    "E741" # Ambiguous variable name
]
lint.select = [
    "C",
    "E",
    "F",
    "I",
    "W",
    "NPY201", # numpy 2.0.0
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
  "F401" # module imported but unused
]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = [
  "eeve"
]